Adding necessary files for version 4.2
This is one of many necessary files for version 4.2 of my web portfolio site. The site itself is derived from a pre-made, MIT-licensed Bulma template (which can be found gere: https://bulmatemplates.github.io/bulma-templates/) so, in a way, I'm "borrowing" it until I finish implementing the other custom-made versions of the site.